Move exceptions2exit from utils.py to exceptions.py
